Deleted serialization project and serializers moved to SQL event store project
