+++ src/EventStores/Domainr.EventStores.sln
@@ -13,10 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mkss.Domainr.EventStore.Ora
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{5F9A7471-2909-4BDC-BA68-9AEB829975D8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mkss.Domainr.EventStore.Serializers", "Mkss.Domainr.EventStore.Serializers\Mkss.Domainr.EventStore.Serializers.csproj", "{E38CC0D1-6ED5-4F72-AC49-52A07B75DF63}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mkss.Domainr.EventStore.Serializers.Tests", "Mkss.Domainr.EventStore.Serializers.Tests\Mkss.Domainr.EventStore.Serializers.Tests.csproj", "{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mkss.Domainr.EventStore.Sqlite", "Mkss.Domainr.EventStore.Sqlite\Mkss.Domainr.EventStore.Sqlite.csproj", "{60ABEB1F-2527-4C45-87C7-D6AA5F2955DF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mkss.Domainr.EventStore.SqlServer.Tests", "Mkss.Domainr.EventStore.SqlServer.Tests\Mkss.Domainr.EventStore.SqlServer.Tests.csproj", "{4B10E32A-6B43-45C3-B1C9-82CF39B6FBEC}"
@@ -36,8 +32,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mkss.Domainr.EventStore.SqlServer", "Mkss.Domainr.EventStore.SqlServer\Mkss.Domainr.EventStore.SqlServer.csproj", "{6E69944B-1378-4A81-8A2F-EB1396512F26}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mkss.Domainr.EventStore.Sql.DependencyInjection", "Mkss.Domainr.EventStore.Sql.DependencyInjection\Mkss.Domainr.EventStore.Sql.DependencyInjection.csproj", "{95B8D411-97DF-42E3-82D9-1F78B514D418}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mkss.Domainr.EventStore.Serializers.DependencyInjection", "Mkss.Domainr.EventStore.Serializers.DependencyInjection\Mkss.Domainr.EventStore.Serializers.DependencyInjection.csproj", "{70814807-23C2-4417-8791-51FA3CEF6E9A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -61,14 +55,6 @@ Global
 		{4E6B675C-3845-40C2-ADA2-9C7128FBAC67}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4E6B675C-3845-40C2-ADA2-9C7128FBAC67}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4E6B675C-3845-40C2-ADA2-9C7128FBAC67}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E38CC0D1-6ED5-4F72-AC49-52A07B75DF63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E38CC0D1-6ED5-4F72-AC49-52A07B75DF63}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E38CC0D1-6ED5-4F72-AC49-52A07B75DF63}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E38CC0D1-6ED5-4F72-AC49-52A07B75DF63}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{60ABEB1F-2527-4C45-87C7-D6AA5F2955DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{60ABEB1F-2527-4C45-87C7-D6AA5F2955DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{60ABEB1F-2527-4C45-87C7-D6AA5F2955DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -109,16 +95,11 @@ Global
 		{95B8D411-97DF-42E3-82D9-1F78B514D418}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{95B8D411-97DF-42E3-82D9-1F78B514D418}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{95B8D411-97DF-42E3-82D9-1F78B514D418}.Release|Any CPU.Build.0 = Release|Any CPU
-		{70814807-23C2-4417-8791-51FA3CEF6E9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{70814807-23C2-4417-8791-51FA3CEF6E9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{70814807-23C2-4417-8791-51FA3CEF6E9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{70814807-23C2-4417-8791-51FA3CEF6E9A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{5E3DF35A-1C4A-459F-B95C-0E9EA55351A7} = {5F9A7471-2909-4BDC-BA68-9AEB829975D8}
 		{4B10E32A-6B43-45C3-B1C9-82CF39B6FBEC} = {5F9A7471-2909-4BDC-BA68-9AEB829975D8}
 		{55877F0D-C094-411F-8BCF-48E66E40CE81} = {5F9A7471-2909-4BDC-BA68-9AEB829975D8}
 		{81269C19-E6CB-4C62-9A18-BE639BE6B5B1} = {5F9A7471-2909-4BDC-BA68-9AEB829975D8}
