+++ src/EventStores/Mkss.Domainr.EventStore.InMemory/Mkss.Domainr.EventStore.InMemory.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
   </ItemGroup>
 
 </Project>
