+++ src/EventStores/Mkss.Domainr.EventStore.Mongo/Mkss.Domainr.EventStore.Mongo.csproj
@@ -7,7 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.7" />
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
     <PackageReference Include="MongoDB.Driver" Version="2.11.0" />
   </ItemGroup>
