+++ src/EventStores/Mkss.Domainr.EventStore.MySql/Mkss.Domainr.EventStore.MySql.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
     <PackageReference Include="MySql.Data" Version="8.0.21" />
   </ItemGroup>
