+++ src/EventStores/Mkss.Domainr.EventStore.Postgresql/Mkss.Domainr.EventStore.Postgresql.csproj
@@ -7,7 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.7" />
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
     <PackageReference Include="Npgsql" Version="4.1.4" />
   </ItemGroup>
