--- src/EventStores/Mkss.Domainr.EventStore.Serializers.DependencyInjection/Mkss.Domainr.EventStore.Serializers.DependencyInjection.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.12" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Mkss.Domainr.EventStore.Serializers\Mkss.Domainr.EventStore.Serializers.csproj" />
-    </ItemGroup>
-
-</Project>
