--- src/EventStores/Mkss.Domainr.EventStore.Serializers/Mkss.Domainr.EventStore.Serializers.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>Domainr.EventStore.Serializers</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-  </ItemGroup>
-
-</Project>
