+++ src/EventStores/Mkss.Domainr.EventStore.Sql.Tests/IntegrationTests/SqlEventStoreTests.cs
@@ -5,10 +5,10 @@ using System.Data.SqlClient;
 using System.Threading.Tasks;
 using Dapper;
 using Domainr.Core.EventSourcing.Abstraction;
-using Domainr.EventStore.Serializers;
 using Domainr.EventStore.Sql.Configuration;
 using Domainr.EventStore.Sql.Data;
 using Domainr.EventStore.Sql.Factories;
+using Domainr.EventStore.Sql.Serializers;
 using Domainr.EventStore.Sql.Tests.Doubles;
 using FluentAssertions;
 using Moq;
