+++ src/EventStores/Mkss.Domainr.EventStore.Sql/Mkss.Domainr.EventStore.Sql.csproj
@@ -8,8 +8,8 @@
   <ItemGroup>
     <PackageReference Include="Dapper" Version="2.0.35" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.7" />
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
-    <PackageReference Include="Mkss.Domainr.EventStore.Serializers" Version="1.0.4" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="System.Runtime.Caching" Version="4.7.0" />
     <PackageReference Include="System.Text.Json" Version="4.7.2" />
   </ItemGroup>
