+++ src/EventStores/Mkss.Domainr.EventStore.SqlServer/Mkss.Domainr.EventStore.SqlServer.csproj
@@ -8,7 +8,7 @@
   <ItemGroup>
     <PackageReference Include="Dapper" Version="2.0.35" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.7" />
-    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.29" />
+    <PackageReference Include="Mkss.Domainr.Core" Version="1.0.30" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.2" />
     <PackageReference Include="System.Runtime.Caching" Version="4.7.0" />
   </ItemGroup>
